+++ css/_variables.scss
@@ -1,0 +1,21 @@
+/*-- COLORS --*/
+$bgMain: #343a40;
+$bgMainRGB: 52,58,64;
+$bgMainRGB_T: rgba(52,58,64,.4);
+$bgMainEndGradientH: #4a4a4a;
+$bgMainEndGradient: 74,74,74;
+$bgOverlay: #2b2b36;
+$bgOverlayRGB: 43,43,54;
+$bgHover: #292e33;
+$bgQuoteRGBA: 52,58,64;
+$bgQuoteRGBA_T: rgba(52,58,64,.8);
+$accent: white;  
+$lightText: #2A2A2A;
+$tableHoverColor: #a31d1d;
+$tableHoverthead:#5c5c5c;
+$btnBg: #7a0c0c;
+$btnBgLight: #d62f2f;
+
+
+/*-- BREAKPOINTS --*/
+$homeBp1: 991px;
