--- js/MVC/Model/local_view_model.js
@@ -1,4 +1,0 @@
-//Local instance of the team data
-var appViewModel={
-  viewModel: teamViewModel
-}
