--- js/MVC/Model/team_data.js
@@ -1,212 +1,0 @@
-//THE TEAMS
-//To keep in theme with the insanity, each team has it's own extreamness
-let theTeams =[
-  {
-    id:0,
-    "Name": "Greasy Humming Birds",
-    "State":"FL",
-    "Coach": "Dianne Jolyon",
-    "MVP": "Sammael Sweetglow",
-    "RemainingPlayers":"1",
-    "Penalties": "3",
-    "TotalHomeruns": "950",
-    "MissilesThwarted": "350",
-    "DamageDealt": "90",
-    "TeamLogo": "img/Team_img/Greasy_Humming_Birds.png"
-  },
-  {
-    id:1,
-		"Name": "Active Stars",
-		"State": "RI",
-    "Coach": "Klotho Archelaos",
-    "MVP": "Theophylaktos Bárðr",
-		"RemainingPlayers": "36",
-		"Penalties": "7",
-		"TotalHomeruns": "13",
-		"MissilesThwarted": "35",
-		"DamageDealt": "13",
-    "TeamLogo": "img/Team_img/Active_Stars.png"
-	},
-	{
-    id:2,
-		"Name": "Exalted Waves",
-		"State": "ID",
-    "Coach": "Deemer Cayley",
-    "MVP": "Kharless Kolynn",
-		"RemainingPlayers": "47",
-		"Penalties": "18",
-		"TotalHomeruns": "70",
-		"MissilesThwarted": "34",
-		"DamageDealt": "60",
-    "TeamLogo": "img/Team_img/Exalted_Waves.png"
-	},
-	{
-    id:3,
-		"Name": "Polar Dawgs",
-		"State": "ND",
-    "Coach": "Shelia Kinsey",
-    "MVP": "ArcaneFrost Overkill",
-		"RemainingPlayers": "26",
-		"Penalties": "19",
-		"TotalHomeruns": "67",
-		"MissilesThwarted": "40",
-		"DamageDealt": "39",
-    "TeamLogo": "img/Team_img/Polar_Dawgs.png"
-	},
-	{    
-    id:4,
-		"Name": "Gruesome Koalas",
-		"State": "OH",
-    "Coach": "Wolf Richmal",
-    "MVP": "Reznor Sweetcheeks",
-		"RemainingPlayers": "9",
-		"Penalties": "5",
-		"TotalHomeruns": "18",
-		"MissilesThwarted": "15",
-		"DamageDealt": "48",
-    "TeamLogo": "img/Team_img/Gruesome_Koalas.png"
-	},
-	{
-    id:5,
-		"Name": "Renowned Mavericks",
-		"State": "TN",
-    "Coach": "Cody Ruben",
-    "MVP": "Phat Hulk",
-		"RemainingPlayers": "33",
-		"Penalties": "2",
-		"TotalHomeruns": "74",
-		"MissilesThwarted": "39",
-		"DamageDealt": "88",
-    "TeamLogo": "img/Team_img/Renowned_Mavericks.png"
-	},
-	{
-    id:6,
-		"Name": "Venemous Gamers",
-		"State": "UT",
-    "Coach": "Mark Fischbach",
-    "MVP": "Mark's Underpaid Editor",
-		"RemainingPlayers": "37",
-		"Penalties": "17",
-		"TotalHomeruns": "50",
-		"MissilesThwarted": "5",
-		"DamageDealt": "16",
-    "TeamLogo": "img/Team_img/Venemous_Gamers.png"
-	},
-	{
-    id:7,
-		"Name": "Witty Tadpols",
-		"State": "NE",
-    "Coach": "Blossom Alphonzo",
-    "MVP": "Luboš Mor",
-		"RemainingPlayers": "22",
-		"Penalties": "9",
-		"TotalHomeruns": "11",
-		"MissilesThwarted": "29",
-		"DamageDealt": "73",
-    "TeamLogo": "img/Team_img/Witty_Tadpols.png"
-	},
-	{
-    id:8,
-		"Name": "Hustlin' Dancers",
-		"State": "MS",
-    "Coach": "Dina Cheftzi-Bah",
-    "MVP": "Llorenç Elektra",
-		"RemainingPlayers": "53",
-		"Penalties": "11",
-		"TotalHomeruns": "54",
-		"MissilesThwarted": "42",
-		"DamageDealt": "23",
-    "TeamLogo": "img/Team_img/Hustling_Dancers.png"
-	},
-	{
-    id:9,
-		"Name": "Skeleton Nets",
-		"State": "NY",
-    "Coach": "Cutthroat Justin The Snooty",
-    "MVP": "First Mate Norma The Smelly",
-		"RemainingPlayers": "62",
-		"Penalties": "2",
-		"TotalHomeruns": "48",
-		"MissilesThwarted": "44",
-		"DamageDealt": "71",
-    "TeamLogo": "img/Team_img/Skeleton_Nets.png"
-	},
-	{
-    id:10,
-		"Name": "Baby Hawks",
-		"State": "SD",
-    "Coach": "Nanny Nolan",
-    "MVP": "Anton Bróðir",
-		"RemainingPlayers": "16",
-		"Penalties": "17",
-		"TotalHomeruns": "24",
-		"MissilesThwarted": "49",
-		"DamageDealt": "44",
-    "TeamLogo": "img/Team_img/Baby_Hawks.png"
-	},
-	{
-    id:11,
-		"Name": "Psychodelic Eagles",
-		"State": "NC",
-    "Coach": "Æthelnoð Kemuel",
-    "MVP": "Sha'ul Yehu'",
-		"RemainingPlayers": "58",
-		"Penalties": "15",
-		"TotalHomeruns": "23",
-		"MissilesThwarted": "16",
-		"DamageDealt": "6",
-    "TeamLogo": "img/Team_img/Psychodelic_Eagles.png"
-	},
-	{
-    id:12,
-		"Name": "Brisk Bullets",
-		"State": "MN",
-    "Coach": "Samson Jamison",
-    "MVP": "Wallis Curtis",
-		"RemainingPlayers": "48",
-		"Penalties": "1",
-		"TotalHomeruns": "61",
-		"MissilesThwarted": "15",
-		"DamageDealt": "19",
-    "TeamLogo": "img/Team_img/Brisk_Bullets.png"
-	},
-	{
-    id:13,
-		"Name": "Lady Lions",
-		"State": "OK",
-    "Coach": "Camellia Sissy",
-    "MVP": "Trent Anabelle",
-		"RemainingPlayers": "46",
-		"Penalties": "13",
-		"TotalHomeruns": "37",
-		"MissilesThwarted": "42",
-		"DamageDealt": "16",
-    "TeamLogo": "img/Team_img/Lady_Lions.png"
-	},
-	{
-    id:14,
-		"Name": "Naughty Raptors",
-		"State": "CA",
-    "Coach": "Ernestine Abraam",
-    "MVP": "Xeno Albanus",
-		"RemainingPlayers": "47",
-		"Penalties": "24",
-		"TotalHomeruns": "13",
-		"MissilesThwarted": "49",
-		"DamageDealt": "39",
-    "TeamLogo": "img/Team_img/Naughty_Raptors.png"
-	},
-	{
-    id:15,
-		"Name": "Holy Demons",
-		"State": "ID",
-    "Coach": "Luvenia Shemu'el",
-    "MVP": "Guðfriðr Sraosha",
-		"RemainingPlayers": "44",
-		"Penalties": "14",
-		"TotalHomeruns": "44",
-		"MissilesThwarted": "47",
-		"DamageDealt": "65",
-    "TeamLogo": "img/Team_img/Holy_Demons.png"
-	}
-]
