--- js/MVC/View/ejs/table_view.ejs
@@ -1,30 +1,0 @@
-<table id='<%=view.remaining.tableID%>' class='<%=view.remaining.tableClasses%>'>
-  <thead>
-    <tr>
-      <% view.view.remaining.tableHead.forEach(function(col){ %>
-        <th scope="col" colspan="<%=col.span%>" class="sortable thead <%=col.otherClasses%>" data-name="<%=col.name%>">
-          <%=col.text%>
-          <i id="<%=col.name%>-asc" class="fa-solid fa-angle-up" aria-hidden="true" style="display:none"></i>
-          <i id="<%=col.name%>-desc" class="fa-solid fa-angle-down" aria-hidden="true" style="display:none"></i>
-        </th>
-      <%})%>
-        <th scope="col" colspan="1" class="text-center" data-name="buttons">Edit</th>>
-    </tr>
-  </thead>
-  <tbody>
-    <% data.forEach(function(d){%>      
-      <tr class="text-center" data-id="<%=d.id%>" data-name='<%=d["Name"]%>' data-bs-toggle="popover">
-        <th><%=d["State"]%></th>
-        <%view.remaining.tableHead.forEach(function(ce){%>
-          <th><%=d[ce.name]%></th>
-        <%})%>
-        <th class="edit-buttons">
-          <!--Edit Buttons-->
-          <span class="fa-solid fa-pencil"></span> 
-          <span class="fa-solid fa-face-frown" data-bs-toggle="modal" data-bs-target="#teamModal" data-name="<%=d['Name']%>"></span>
-        </th>
-      </tr>
-    <%})%>
-  </tbody>  
-</table>
-
