+++ js/Util/util.js
@@ -1,0 +1,12 @@
+export default class Utilities{
+    constructor()
+    {
+       this.files={}
+    }
+    async getFileContents(url){
+        if (!(url in this.files)){
+            this.files[url]=await $.get(url);
+        }       
+        return this.files[url]        
+     }
+}
