+++ js/View/View_Model/local_view_model.js
@@ -1,0 +1,6 @@
+//Local instance of the team data
+import teamViewModel from './team_view_model.js'
+var appViewModel={
+  viewModel: teamViewModel
+}
+export default appViewModel;
