+++ js/View/ejs/Modal/delTeam_modal.ejs
@@ -1,0 +1,18 @@
+  <div class="modal-dialog modal-dialog-centered">
+    <div class="modal-content">
+      <div class="modal-header">
+        <img class="img-fluid rounded-circle modal-pic" src='<%=logo%>' width="50" height="50">
+        <h5 class="modal-title w-100">Banish The <%=name%> ?</h2>
+        <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
+      </div>
+      
+      <div class="modal-body">
+        This will temporarily send them to the shadow dimension and you will be held responsible in a court of law for any potential psycological or meta-spritual damages. Also, <%=mvp%> will be really sad
+      </div>
+      
+      <div class="modal-footer">
+        <button type="button" class="btn super-btn-no" data-bs-dismiss="modal">What? No!</button>
+        <button type="button" class="btn super-btn" id="banish-button" data-bs-dismiss="modal">They Know What They Did</button>
+      </div>
+    </div>
+  </div>
